fix(index): remove options as it has to be provided over options plugin since webpack >=2.1.0-beta.23 (#3)
BREAKING CHANGE: removed 'options' as it has to be provided over options plugin from webpack >=2.1.0-beta.23.

Note: We can use webpack.LoaderOptionsPlugin only once, so if we use this in the plugin directly we can't use it again. So just removing the tslint from the root config of webpack in this plugin and updating somewhere in the readme should be fine.